pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chathrd"
version = "0.1.0"
description = "Система интеллектуального поиска по HR документам"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "chromadb>=0.4.18",
    "pydantic>=2.5.2",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.3.1",
    "unstructured>=0.11.0",
    "unstructured-inference>=0.7.18",
    "pdf2image>=1.16.3",
    "pillow>=10.1.0",
    "pypdf>=3.17.0",
    "docx2txt>=0.8",
    "python-pptx>=0.6.21",
    "pytesseract>=0.3.10",
    "openai>=1.1.1",
    "python-telegram-bot>=20.7",
    "typer>=0.9.0",
    "loguru>=0.7.0",
    "rich>=13.6.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.10.1",
    "isort>=5.12.0",
    "mypy>=1.6.1",
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.3",
]
notebooks = [
    "ipykernel>=6.26.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.8.1",
    "pandas>=2.1.2",
    "scikit-learn>=1.3.2",
    "seaborn>=0.13.0",
]

[project.scripts]
chathrd-index = "chathrd.cli.index_command:main"
chathrd-query = "chathrd.cli.query_command:main"
# Точка входа для Telegram бота была удалена, так как бот теперь запускается из telegram_bot/bot.py

[tool.hatch.build.targets.wheel]
packages = ["src/chathrd"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "B", "W", "I", "N", "UP", "COM", "C4", "EXE", "ISC", "PIE", "T20", "PT", "ERA"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*" 